Rename from artemis to generic server. Add notificationType to body.
