+++ src/main/resources/english.properties
@@ -1,3 +1,1 @@
-artemis.webhookUrl = Server URL
-artemis.webhookUrl.description = Your Server incoming URL.
-artemis.webhookUrl.error = You must enter a URL.
+server.webhookUrl = Server URL
