+++ src/main/resources/serverEditNotification.ftl
@@ -1,0 +1,1 @@
+[@ww.textfield labelKey="server.webhookUrl" name="webhookUrl" value="${webhookUrl!}" required='true'/]
