+++ src/main/resources/serverViewNotification.ftl
